using jwt claims and @jwt_optional

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, jwt_optional, get_jwt_claims, get_jwt_identity
 from models.item import ItemModel
 
 
@@ -40,6 +40,10 @@ class Item(Resource):
 
     @jwt_required
     def put(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message": "Admin privilege required"}, 401
+
         data = Item.parser.parse_args()
 
         item = ItemModel.find_by_name(name)
@@ -55,6 +59,10 @@ class Item(Resource):
 
     @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message": "Admin privilege required"}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -62,5 +70,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {"items": [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items": items}, 200
+        return {
+            "items": [item['name'] for item in items],
+            "message": "Login to learn more about these items"
+        }, 200
